requirements.txt
requests
pandas
amqp==2.3.2
beautifulsoup4==4.6.0
billiard==3.5.0.3
celery==4.2.0
click==6.7
Flask==1.0.2
Flask-PyMongo==0.5.2
Flask-Session==0.3.1
gunicorn==19.8.1
itsdangerous==0.24
Jinja2==2.10
kombu==4.2.1
MarkupSafe==1.0
nltk==3.3
peewee==3.5.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycode-similar==1.2
pymongo==3.6.1
python-Levenshtein==0.12.0
pytz==2018.4
rake-nltk==1.0.4
redis==2.10.6
six==1.11.0
vine==1.1.4
Werkzeug==0.14.1
